Move Attach methods to context
To match EF
