+++ src/MongoFramework/IMongoDbSet.cs
@@ -23,8 +23,6 @@ namespace MongoFramework
 		TEntity Create();
 		void Add(TEntity entity);
 		void AddRange(IEnumerable<TEntity> entities);
-		void Attach(TEntity entity);
-		void AttachRange(IEnumerable<TEntity> entities);
 		void Update(TEntity entity);
 		void UpdateRange(IEnumerable<TEntity> entities);
 		void Remove(TEntity entity);
