+++ tests/MongoFramework.Tests/MongoDbContextTenantTests.cs
@@ -111,5 +111,133 @@ namespace MongoFramework.Tests
 				Assert.AreEqual(TestConfiguration.GetTenantId(), context.DbSet.First().TenantId);
 			}
 		}
+		
+				[TestMethod]
+		public void SuccessfullyAttachUntrackedEntity()
+		{
+			var connection = TestConfiguration.GetConnection();
+			var tenantId = TestConfiguration.GetTenantId();
+			var context = new MongoDbTenantContext(connection, tenantId);
+			var dbSet = new MongoDbTenantSet<DbSetModel>(context);
+
+			var model = new DbSetModel
+			{
+				Id = "abcd"
+			};
+
+			dbSet.Add(model);
+
+			context.SaveChanges();
+
+			ResetMongoDb();
+
+			context = new MongoDbTenantContext(connection, tenantId);
+			dbSet = new MongoDbTenantSet<DbSetModel>(context);
+
+			var result = dbSet.AsNoTracking().FirstOrDefault();
+
+			context.Attach(result);
+
+			Assert.AreEqual(MongoFramework.Infrastructure.EntityEntryState.NoChanges, context.ChangeTracker.GetEntry(result).State);
+		}
+
+		[TestMethod]
+		public void SuccessfullyAttachUntrackedEntities()
+		{
+			var connection = TestConfiguration.GetConnection();
+			var tenantId = TestConfiguration.GetTenantId();
+			var context = new MongoDbTenantContext(connection, tenantId);
+			var dbSet = new MongoDbTenantSet<DbSetModel>(context);
+
+			var entities = new[] {
+				new DbSetModel
+				{
+					Id = "abcd"
+				},
+				new DbSetModel
+				{
+					Id = "efgh"
+				}
+			};
+
+			dbSet.AddRange(entities);
+
+			context.SaveChanges();
+
+			ResetMongoDb();
+
+			context = new MongoDbTenantContext(connection, tenantId);
+			dbSet = new MongoDbTenantSet<DbSetModel>(context);
+
+			var result = dbSet.AsNoTracking().ToList();
+
+			context.AttachRange(result);
+
+			Assert.AreEqual(MongoFramework.Infrastructure.EntityEntryState.NoChanges, context.ChangeTracker.GetEntry(result[0]).State);
+			Assert.AreEqual(MongoFramework.Infrastructure.EntityEntryState.NoChanges, context.ChangeTracker.GetEntry(result[1]).State);
+		}
+
+		[TestMethod]
+		public void AttachRejectsMismatchedEntity()
+		{
+			var connection = TestConfiguration.GetConnection();
+			var tenantId = TestConfiguration.GetTenantId();
+			var context = new MongoDbTenantContext(connection, tenantId);
+			var dbSet = new MongoDbTenantSet<DbSetModel>(context);
+
+			var model = new DbSetModel
+			{
+				Id = "abcd"
+			};
+
+			dbSet.Add(model);
+
+			context.SaveChanges();
+
+			ResetMongoDb();
+
+			context = new MongoDbTenantContext(connection, tenantId);
+			dbSet = new MongoDbTenantSet<DbSetModel>(context);
+
+			var result = dbSet.AsNoTracking().FirstOrDefault();
+			result.TenantId = tenantId + "a";
+
+			Assert.ThrowsException<MultiTenantException>(() => context.Attach(result));
+		}
+
+		[TestMethod]
+		public void AttachRejectsMismatchedEntities()
+		{
+			var connection = TestConfiguration.GetConnection();
+			var tenantId = TestConfiguration.GetTenantId();
+			var context = new MongoDbTenantContext(connection, tenantId);
+			var dbSet = new MongoDbTenantSet<DbSetModel>(context);
+
+			var entities = new[] {
+				new DbSetModel
+				{
+					Id = "abcd"
+				},
+				new DbSetModel
+				{
+					Id = "efgh"
+				}
+			};
+
+			dbSet.AddRange(entities);
+
+			context.SaveChanges();
+
+			ResetMongoDb();
+
+			context = new MongoDbTenantContext(connection, tenantId);
+			dbSet = new MongoDbTenantSet<DbSetModel>(context);
+
+			var result = dbSet.AsNoTracking().ToList();
+			result[0].TenantId = tenantId + "a";
+
+			Assert.ThrowsException<MultiTenantException>(() => context.AttachRange(result));
+		}
+
 	}
 }
