+++ tests/MongoFramework.Tests/MongoDbSetTests.cs
@@ -470,72 +470,7 @@ namespace MongoFramework.Tests
 
 			Assert.IsNull(context.ChangeTracker.GetEntry(result));
 		}
-
-		[TestMethod]
-		public void SuccessfullyAttachUntrackedEntity()
-		{
-			var connection = TestConfiguration.GetConnection();
-			var context = new MongoDbContext(connection);
-			var dbSet = new MongoDbSet<TestModel>(context);
-
-			var model = new TestModel
-			{
-				Id = "abcd",
-				Description = "SuccessfullyAttachUntrackedEntity.1"
-			};
-
-			dbSet.Add(model);
-
-			context.SaveChanges();
-
-			ResetMongoDb();
-
-			context = new MongoDbContext(connection);
-			dbSet = new MongoDbSet<TestModel>(context);
-
-			var result = dbSet.AsNoTracking().FirstOrDefault();
-
-			dbSet.Attach(result);
-
-			Assert.AreEqual(MongoFramework.Infrastructure.EntityEntryState.NoChanges, context.ChangeTracker.GetEntry(result).State);
-		}
-
-		[TestMethod]
-		public void SuccessfullyAttachUntrackedEntities()
-		{
-			var connection = TestConfiguration.GetConnection();
-			var context = new MongoDbContext(connection);
-			var dbSet = new MongoDbSet<TestModel>(context);
-
-			var entities = new[] {
-				new TestModel
-				{
-					Description = "SuccessfullyAttachUntrackedEntities.1"
-				},
-				new TestModel
-				{
-					Description = "SuccessfullyAttachUntrackedEntities.2",
-					BooleanField = true
-				}
-			};
-
-			dbSet.AddRange(entities);
-
-			context.SaveChanges();
-
-			ResetMongoDb();
-
-			context = new MongoDbContext(connection);
-			dbSet = new MongoDbSet<TestModel>(context);
-
-			var result = dbSet.AsNoTracking().ToList();
-
-			dbSet.AttachRange(result);
-
-			Assert.AreEqual(MongoFramework.Infrastructure.EntityEntryState.NoChanges, context.ChangeTracker.GetEntry(result[0]).State);
-			Assert.AreEqual(MongoFramework.Infrastructure.EntityEntryState.NoChanges, context.ChangeTracker.GetEntry(result[1]).State);
-		}
-
+		
 	}
 
 }
