+++ tests/MongoFramework.Tests/MongoDbTenantSetTests.cs
@@ -931,136 +931,5 @@ namespace MongoFramework.Tests
 			Assert.IsNull(context.ChangeTracker.GetEntry(result));		
 		}
 
-		[TestMethod]
-		public void SuccessfullyAttachUntrackedEntity()
-		{
-			var connection = TestConfiguration.GetConnection();
-			var tenantId = TestConfiguration.GetTenantId();
-			var context = new MongoDbTenantContext(connection, tenantId);
-			var dbSet = new MongoDbTenantSet<TestModel>(context);
-
-			var model = new TestModel
-			{
-				Id = "abcd",
-				Description = "SuccessfullyAttachUntrackedEntity.1"
-			};
-
-			dbSet.Add(model);
-
-			context.SaveChanges();
-
-			ResetMongoDb();
-
-			context = new MongoDbTenantContext(connection, tenantId);
-			dbSet = new MongoDbTenantSet<TestModel>(context);
-
-			var result = dbSet.AsNoTracking().FirstOrDefault();
-
-			dbSet.Attach(result);
-
-			Assert.AreEqual(MongoFramework.Infrastructure.EntityEntryState.NoChanges, context.ChangeTracker.GetEntry(result).State);
-		}
-
-		[TestMethod]
-		public void SuccessfullyAttachUntrackedEntities()
-		{
-			var connection = TestConfiguration.GetConnection();
-			var tenantId = TestConfiguration.GetTenantId();
-			var context = new MongoDbTenantContext(connection, tenantId);
-			var dbSet = new MongoDbTenantSet<TestModel>(context);
-
-			var entities = new[] {
-				new TestModel
-				{
-					Description = "SuccessfullyAttachUntrackedEntities.1"
-				},
-				new TestModel
-				{
-					Description = "SuccessfullyAttachUntrackedEntities.2",
-					BooleanField = true
-				}
-			};
-
-			dbSet.AddRange(entities);
-
-			context.SaveChanges();
-
-			ResetMongoDb();
-
-			context = new MongoDbTenantContext(connection, tenantId);
-			dbSet = new MongoDbTenantSet<TestModel>(context);
-
-			var result = dbSet.AsNoTracking().ToList();
-
-			dbSet.AttachRange(result);
-
-			Assert.AreEqual(MongoFramework.Infrastructure.EntityEntryState.NoChanges, context.ChangeTracker.GetEntry(result[0]).State);
-			Assert.AreEqual(MongoFramework.Infrastructure.EntityEntryState.NoChanges, context.ChangeTracker.GetEntry(result[1]).State);
-		}
-
-		[TestMethod]
-		public void AttachRejectsMismatchedEntity()
-		{
-			var connection = TestConfiguration.GetConnection();
-			var tenantId = TestConfiguration.GetTenantId();
-			var context = new MongoDbTenantContext(connection, tenantId);
-			var dbSet = new MongoDbTenantSet<TestModel>(context);
-
-			var model = new TestModel
-			{
-				Id = "abcd",
-				Description = "AttachRejectsMismatchedEntity.1"
-			};
-
-			dbSet.Add(model);
-
-			context.SaveChanges();
-
-			ResetMongoDb();
-
-			context = new MongoDbTenantContext(connection, tenantId);
-			dbSet = new MongoDbTenantSet<TestModel>(context);
-
-			var result = dbSet.AsNoTracking().FirstOrDefault();
-			result.TenantId = tenantId + "a";
-
-			Assert.ThrowsException<MultiTenantException>(() => dbSet.Attach(result));
-		}
-
-		[TestMethod]
-		public void AttachRejectsMismatchedEntities()
-		{
-			var connection = TestConfiguration.GetConnection();
-			var tenantId = TestConfiguration.GetTenantId();
-			var context = new MongoDbTenantContext(connection, tenantId);
-			var dbSet = new MongoDbTenantSet<TestModel>(context);
-
-			var entities = new[] {
-				new TestModel
-				{
-					Description = "AttachRejectsMismatchedEntities.1"
-				},
-				new TestModel
-				{
-					Description = "AttachRejectsMismatchedEntities.2",
-					BooleanField = true
-				}
-			};
-
-			dbSet.AddRange(entities);
-
-			context.SaveChanges();
-
-			ResetMongoDb();
-
-			context = new MongoDbTenantContext(connection, tenantId);
-			dbSet = new MongoDbTenantSet<TestModel>(context);
-
-			var result = dbSet.AsNoTracking().ToList();
-			result[0].TenantId = tenantId + "a";
-
-			Assert.ThrowsException<MultiTenantException>(() => dbSet.AttachRange(result));
-		}
-
 	}
 }
